refactor: use role-based B5-W0 layout
Code co-developed by Theo and Codex
