+++ .gitignore
@@ -104,14 +104,12 @@ research/asset-stewardship/asw-0b5-world-family/*
 !research/asset-stewardship/asw-0b5-world-family/declarations/
 research/asset-stewardship/asw-0b5-world-family/declarations/*
 !research/asset-stewardship/asw-0b5-world-family/declarations/w1-member-authority.json
-!research/asset-stewardship/asw-0b5-world-family/src/
-research/asset-stewardship/asw-0b5-world-family/src/*
-!research/asset-stewardship/asw-0b5-world-family/src/asw_b5_generator_boundary/
-!research/asset-stewardship/asw-0b5-world-family/src/asw_b5_generator_boundary/*.py
-!research/asset-stewardship/asw-0b5-world-family/src/asw_b5_certifier_boundary/
-!research/asset-stewardship/asw-0b5-world-family/src/asw_b5_certifier_boundary/*.py
-!research/asset-stewardship/asw-0b5-world-family/src/asw_b5_lineage/
-!research/asset-stewardship/asw-0b5-world-family/src/asw_b5_lineage/*.py
+!research/asset-stewardship/asw-0b5-world-family/generator/
+!research/asset-stewardship/asw-0b5-world-family/generator/*.py
+!research/asset-stewardship/asw-0b5-world-family/certifier/
+!research/asset-stewardship/asw-0b5-world-family/certifier/*.py
+!research/asset-stewardship/asw-0b5-world-family/lineage/
+!research/asset-stewardship/asw-0b5-world-family/lineage/*.py
 !research/asset-stewardship/asw-0b5-world-family/tests/
 research/asset-stewardship/asw-0b5-world-family/tests/*
 !research/asset-stewardship/asw-0b5-world-family/tests/support.py
